refactor: separate package directories
