+++ nextstep-mvc/src/main/java/nextstep/mvc/DispatcherServlet.java
@@ -1,8 +1,10 @@
 package nextstep.mvc;
 
-import nextstep.mvc.exception.NotFoundHandlerAdapter;
 import nextstep.mvc.tobe.ModelAndView;
-import nextstep.mvc.tobe.View;
+import nextstep.mvc.tobe.adapter.HandlerAdapter;
+import nextstep.mvc.tobe.exception.NotFoundHandlerAdapter;
+import nextstep.mvc.tobe.mapping.HandlerMapping;
+import nextstep.mvc.tobe.view.View;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
