+++ nextstep-mvc/src/main/java/nextstep/mvc/tobe/view/JsonView.java
@@ -1,4 +1,4 @@
-package nextstep.mvc.tobe;
+package nextstep.mvc.tobe.view;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import nextstep.web.support.MediaType;
